fix(tests): Update Q4_0 tests for 18-byte block format
Fixes property and smoke tests to use correct Q4_0 block format:
- Block size: 18 bytes (2-byte f16 scale + 16 quants), not 20 bytes
- Scale: f16 (half precision), not f32

Changes:
- property_quantize.rs: Update strategy and bounds tests for f16 scale
- smoke_e2e.rs: Fix block creation for 18-byte format
- Code formatting (rustfmt) in apr_transformer.rs, quantize.rs, examples

All 792 lib tests + 26 integration tests pass.

+++ examples/convert_and_bench_apr.rs
@@ -18,11 +18,21 @@ fn print_cpu_features() {
 
     println!("   CPU Features:");
     print!("     ");
-    if is_x86_feature_detected!("avx2") { print!("AVX2 "); }
-    if is_x86_feature_detected!("fma") { print!("FMA "); }
-    if is_x86_feature_detected!("avx512f") { print!("AVX-512 "); }
-    if is_x86_feature_detected!("avx512vnni") { print!("AVX512-VNNI "); }
-    if is_x86_feature_detected!("avxvnni") { print!("AVX-VNNI "); }
+    if is_x86_feature_detected!("avx2") {
+        print!("AVX2 ");
+    }
+    if is_x86_feature_detected!("fma") {
+        print!("FMA ");
+    }
+    if is_x86_feature_detected!("avx512f") {
+        print!("AVX-512 ");
+    }
+    if is_x86_feature_detected!("avx512vnni") {
+        print!("AVX512-VNNI ");
+    }
+    if is_x86_feature_detected!("avxvnni") {
+        print!("AVX-VNNI ");
+    }
     println!();
 }
 
@@ -62,7 +72,10 @@ fn main() {
     println!("\n2. Creating APR Q4_0...");
     let start = Instant::now();
     let apr_q4 = QuantizedAprTransformerQ4::from_gguf(&gguf_model);
-    println!("   Conversion completed in {:.2}s", start.elapsed().as_secs_f32());
+    println!(
+        "   Conversion completed in {:.2}s",
+        start.elapsed().as_secs_f32()
+    );
 
     // === Benchmark GGUF (single token, no context) ===
     println!("\n3. Benchmarking GGUF (single token)...");
@@ -77,7 +90,11 @@ fn main() {
     }
     let gguf_avg = start.elapsed() / iterations;
     let gguf_tps = 1.0 / gguf_avg.as_secs_f64();
-    println!("   Throughput: {:.1} tok/s ({:.1}ms/tok)", gguf_tps, gguf_avg.as_secs_f64() * 1000.0);
+    println!(
+        "   Throughput: {:.1} tok/s ({:.1}ms/tok)",
+        gguf_tps,
+        gguf_avg.as_secs_f64() * 1000.0
+    );
 
     // === Benchmark APR Q4_0 (single token, no context) ===
     println!("\n4. Benchmarking APR Q4_0 (single token, parallel)...");
@@ -91,7 +108,11 @@ fn main() {
     }
     let apr_q4_avg = start.elapsed() / iterations;
     let apr_q4_tps = 1.0 / apr_q4_avg.as_secs_f64();
-    println!("   Throughput: {:.1} tok/s ({:.1}ms/tok)", apr_q4_tps, apr_q4_avg.as_secs_f64() * 1000.0);
+    println!(
+        "   Throughput: {:.1} tok/s ({:.1}ms/tok)",
+        apr_q4_tps,
+        apr_q4_avg.as_secs_f64() * 1000.0
+    );
 
     // === Benchmark APR Q4_0 with KV Cache (context-aware generation) ===
     println!("\n5. Benchmarking APR Q4_0 + KV Cache (context-aware)...");
@@ -109,13 +130,21 @@ fn main() {
     let gen_tokens = 20u32;
     let start = Instant::now();
     for i in 0..gen_tokens {
-        let _ = apr_q4.forward_with_cache(&[(i % 100) as u32], &mut cache);
+        let _ = apr_q4.forward_with_cache(&[((i % 100))], &mut cache);
     }
     let cache_total = start.elapsed();
     let cache_avg = cache_total / gen_tokens;
     let cache_tps = gen_tokens as f64 / cache_total.as_secs_f64();
-    println!("   Generated {} tokens in {:.0}ms", gen_tokens, cache_total.as_secs_f64() * 1000.0);
-    println!("   Throughput: {:.1} tok/s ({:.1}ms/tok avg)", cache_tps, cache_avg.as_secs_f64() * 1000.0);
+    println!(
+        "   Generated {} tokens in {:.0}ms",
+        gen_tokens,
+        cache_total.as_secs_f64() * 1000.0
+    );
+    println!(
+        "   Throughput: {:.1} tok/s ({:.1}ms/tok avg)",
+        cache_tps,
+        cache_avg.as_secs_f64() * 1000.0
+    );
 
     // === Benchmark APR Q4_0 with Scratch Buffers (zero-allocation) ===
     println!("\n6. Benchmarking APR Q4_0 + Scratch Buffers (zero-alloc)...");
@@ -127,11 +156,15 @@ fn main() {
     }
     let start = Instant::now();
     for i in 0..iterations {
-        let _ = apr_q4.forward_single_with_scratch((i % 100) as u32, &mut scratch);
+        let _ = apr_q4.forward_single_with_scratch(((i % 100)), &mut scratch);
     }
     let scratch_avg = start.elapsed() / iterations;
     let scratch_tps = 1.0 / scratch_avg.as_secs_f64();
-    println!("   Throughput: {:.1} tok/s ({:.1}ms/tok)", scratch_tps, scratch_avg.as_secs_f64() * 1000.0);
+    println!(
+        "   Throughput: {:.1} tok/s ({:.1}ms/tok)",
+        scratch_tps,
+        scratch_avg.as_secs_f64() * 1000.0
+    );
     println!("   vs allocating: {:.1}x", scratch_tps / apr_q4_tps);
 
     // === Summary ===
@@ -142,32 +175,58 @@ fn main() {
     println!("╠══════════════════════════════════════════════════════════════════════╣");
     println!(
         "║ GGUF Q4_0        │ {:>7.1} tok/s │ {:>6.1}ms │ 1.00x   │ None         ║",
-        gguf_tps, gguf_avg.as_secs_f64() * 1000.0
+        gguf_tps,
+        gguf_avg.as_secs_f64() * 1000.0
     );
     println!(
         "║ APR Q4_0         │ {:>7.1} tok/s │ {:>6.1}ms │ {:.2}x   │ None         ║",
-        apr_q4_tps, apr_q4_avg.as_secs_f64() * 1000.0, apr_q4_tps / gguf_tps
+        apr_q4_tps,
+        apr_q4_avg.as_secs_f64() * 1000.0,
+        apr_q4_tps / gguf_tps
     );
     println!(
         "║ APR + KV Cache   │ {:>7.1} tok/s │ {:>6.1}ms │ {:.2}x   │ {} tokens   ║",
-        cache_tps, cache_avg.as_secs_f64() * 1000.0, cache_tps / gguf_tps, gen_tokens
+        cache_tps,
+        cache_avg.as_secs_f64() * 1000.0,
+        cache_tps / gguf_tps,
+        gen_tokens
     );
     println!(
         "║ APR + Scratch    │ {:>7.1} tok/s │ {:>6.1}ms │ {:.2}x   │ Zero-alloc   ║",
-        scratch_tps, scratch_avg.as_secs_f64() * 1000.0, scratch_tps / gguf_tps
+        scratch_tps,
+        scratch_avg.as_secs_f64() * 1000.0,
+        scratch_tps / gguf_tps
     );
     println!("╚══════════════════════════════════════════════════════════════════════╝");
 
     // Analysis
     println!("\nKey Insights:");
-    println!("  • Single-token (allocating): APR is {:.1}x faster than GGUF", apr_q4_tps / gguf_tps);
-    println!("  • Scratch buffers: {:.1}x speedup over allocating path", scratch_tps / apr_q4_tps);
-    println!("  • Context-aware generation: {:.1} tok/s with KV cache", cache_tps);
+    println!(
+        "  • Single-token (allocating): APR is {:.1}x faster than GGUF",
+        apr_q4_tps / gguf_tps
+    );
+    println!(
+        "  • Scratch buffers: {:.1}x speedup over allocating path",
+        scratch_tps / apr_q4_tps
+    );
+    println!(
+        "  • Context-aware generation: {:.1} tok/s with KV cache",
+        cache_tps
+    );
     println!("  • KV cache enables attending to past tokens efficiently");
 
     let llama_cpp_tps = 42.0;
     println!("\n  vs llama.cpp (42 tok/s):");
-    println!("    • APR (allocating):   {:>5.0}%", apr_q4_tps / llama_cpp_tps * 100.0);
-    println!("    • APR + Scratch:      {:>5.0}%", scratch_tps / llama_cpp_tps * 100.0);
-    println!("    • APR + KV Cache:     {:>5.0}%", cache_tps / llama_cpp_tps * 100.0);
+    println!(
+        "    • APR (allocating):   {:>5.0}%",
+        apr_q4_tps / llama_cpp_tps * 100.0
+    );
+    println!(
+        "    • APR + Scratch:      {:>5.0}%",
+        scratch_tps / llama_cpp_tps * 100.0
+    );
+    println!(
+        "    • APR + KV Cache:     {:>5.0}%",
+        cache_tps / llama_cpp_tps * 100.0
+    );
 }
